Its purpose
Basic Text edit game in which you begin in a forest talking to a goblin. 
He will ask you for a favor to obtain ingrediants that were stolen from giants. 
The giants fled into a castle.
Your mission, is to go through the 5 floors, battling through with enemies, level up, 
and reach the boss room where the giant is located. 
Defeat the giant, obtain the ingrediants, return to the goblin, game complete. 

More will be added soon.